moved personalAccessKeyPrompt to cms-cli/lib/prompts.js -- causes cyclic dependency issue
