+++ packages/cms-cli/package.json
@@ -12,6 +12,7 @@
     "@hubspot/cms-lib": "^1.1.3-beta.1",
     "commander": "^2.16.0",
     "inquirer": "^6.3.1",
+    "open": "^7.0.3",
     "ora": "^4.0.3",
     "readline": "^1.3.0",
     "shelljs": "0.8.3",
